requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
openpyxl==3.1.5
reportlab==4.0.4
pytz==2024.1
Werkzeug==2.3.8
